+++ DearPyGui/src/core/AppItems/plots/mvCandleSeries.cpp
@@ -30,7 +30,7 @@ namespace Marvel {
         // get ImGui window DrawList
         ImDrawList* draw_list = ImPlot::GetPlotDrawList();
         // calc real value width
-        float half_width = count > 1 ? (xs[1] - xs[0]) * width_percent : width_percent;
+        float half_width = count > 1 ? ((float)xs[1] - (float)xs[0]) * width_percent : width_percent;
 
         // custom tool
         if (ImPlot::IsPlotHovered() && tooltip) {
